# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Puspa API Documentation",
        "description": "Dokumentasi API lengkap untuk layanan Puspa",
        "contact": {
            "email": "admin@puspa.sinus.ac.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Registrasi pengguna baru",
                "description": "Mendaftarkan pengguna baru dan mengirimkan email verifikasi.",
                "operationId": "registerUser",
                "requestBody": {
                    "description": "Data registrasi pengguna",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Registrasi berhasil. Silakan melakukan verifikasi!"
                                        },
                                        "data": {
                                            "properties": {
                                                "user_id": {
                                                    "description": "ULID dari pengguna yang baru dibuat",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login pengguna",
                "description": "Mengautentikasi pengguna dan mengembalikan access token.",
                "operationId": "loginUser",
                "requestBody": {
                    "description": "Kredensial login",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login berhasil"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/LoginResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Kredensial tidak valid"
                    },
                    "403": {
                        "description": "Akun belum aktif atau belum diverifikasi"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout pengguna",
                "description": "Mencabut (revoke) token pengguna yang sedang login.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/protected": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Protected route",
                "description": "Protected route example.",
                "operationId": "7cd02503e789fc0346399eef296a1076",
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Kirim link reset password",
                "description": "Mengirim email berisi link untuk mereset password pengguna.",
                "operationId": "forgotPassword",
                "requestBody": {
                    "description": "Email pengguna yang terdaftar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john.doe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Link berhasil dikirim"
                    },
                    "404": {
                        "description": "Pengguna dengan email tersebut tidak ditemukan"
                    },
                    "422": {
                        "description": "Email tidak valid"
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset password pengguna",
                "description": "Mengatur ulang password pengguna menggunakan token dari email.",
                "operationId": "resetPassword",
                "requestBody": {
                    "description": "Data untuk reset password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResetPasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password berhasil diatur ulang"
                    },
                    "422": {
                        "description": "Data tidak valid (misal: token salah, email tidak cocok, atau validasi password gagal)"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ResetPasswordRequest": {
                "required": [
                    "token",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "token": {
                        "description": "Token yang diterima dari email",
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "NewPassword123!"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "NewPassword123!"
                    }
                },
                "type": "object"
            },
...
            "UserLoginRequest": {
                "required": [
                    "identifier",
                    "password"
                ],
                "properties": {
                    "identifier": {
                        "description": "Bisa berupa username atau email",
                        "type": "string",
                        "example": "johndoe"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Password123!"
                    }
                },
                "type": "object"
            },
            "UserRegisterRequest": {
                "required": [
                    "username",
                    "email",
                    "password"
                ],
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "description": "Min. 8 karakter, 1 huruf besar, 1 angka, 1 simbol",
                        "type": "string",
                        "format": "password",
                        "example": "Password123!"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "Password123!"
                    }
                },
                "type": "object"
            },
            "LoginResource": {
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "string"
                    },
                    "username": {
                        "description": "Username",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email address",
                        "type": "string",
                        "format": "email"
                    },
                    "role": {
                        "description": "User role",
                        "type": "string"
                    },
                    "token_type": {
                        "type": "string",
                        "example": "Bearer"
                    },
                    "access_token": {
                        "description": "Authentication token",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token Bearer Anda",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}